Step4: Tested the db access and graph traversal
Queries and traversal on the graph work with all the pieces being in
place. Things needed to get it working

1. Cleaned up email addresses (containing spaces or two consecutive
dots)

2. Installed apoc for neo4j: The shortest path using the built in search
mechanism in neo4j was taking extraordinarily long times, even for small
paths. This was solved using the dijkstra search in the apoc plugin.

3. Tested graph search with apoc.algo.dijkstra. Works well and tested
this for the example specified - Lynn Blair and tidwellgreer@bfusa.com.
This query returns a rather long path between the two entities.

4. While this query works, this is unacceptably slow for a REST API.
The next step is thus to speed things up.

5. In order to speed things up, two things will be implemented-
a. A smaller model which gets rid of the Recipient nodes
b. If needed, computing the all pairs shortest path for all email
addresses and storing it in the cache. Since we have about 75000 unique
email addresses, this will required about 560 million entries. However,
that should easily achievable using either a high end machine or a small
sized cluster. Either way, this step will be done only if necessary.

6. Still TODO

a. Moving this to docker and automating installation/migration.