Step5: changed the data model
Neo4j was suffering immensely under the old data model. Changed it
significantly by processing the data in the following ways

1. Normalized the email addresses out into a separate table and created
links with the employee table where they were called for.

2. Removed all the nodes pertaining to Message and RecipientInfo
and created direct links between the Email addresses if the first email
address has sent an email to the second one.

3. Merged all instances of emails sent between two nodes into one.

4. Removed self-emails. These loops were showing up in the Neo4j
returned path. This shouldn't happen as the node should be marked
visited on first entry. This suggests that the implementation of
Dijktra's SPF has a few bugs. Removing self-emails solves this problem
for now.

5. With the previous change, the code works quite fast. Will test it out
for other scenarios and also make sure that it handles failures
gracefully, besides integrating with Flask.